fix 4details
